feat: Add NavigationManager
Navigation 디자인에 맞게 변경, 아이콘 추가
Navigation 관리 페이지 추가 및 이동
